move coda specific logic to l10n_be_coda_advanced
