+++ account_bank_statement_advanced/models/account_bank_statement.py
@@ -1,9 +1,8 @@
 # -*- coding: utf-8 -*-
-# Copyright 2009-2018 Noviat.
+# Copyright 2009-2019 Noviat.
 # License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl).
 
 from lxml import etree
-import time
 
 from odoo import api, models, _
 
@@ -55,31 +54,6 @@ class AccountBankStatement(models.Model):
         cancel statements while preserving associated moves.
         """
         self.state = 'open'
-        return True
-
-    @api.multi
-    def button_confirm_bank(self):
-        """
-        Some of the function in this method are handled (differently) by the
-        CODA processing hence we bypass those here.
-        """
-        coda_statements = self.filtered(lambda r: r.coda_id)
-        non_coda_statements = self - coda_statements
-        super(AccountBankStatement, non_coda_statements).button_confirm_bank()
-
-        coda_statements._balance_check()
-        coda_statements = coda_statements.filtered(
-            lambda r: r.state == 'open')
-        moves = coda_statements.mapped('line_ids').mapped('journal_entry_ids')
-        unposted = moves.filtered(lambda r: r.state == 'draft')
-        unposted.post()
-        for statement in coda_statements:
-            statement.message_post(
-                body=_('Statement %s confirmed, journal items were created.'
-                       ) % (statement.name,))
-        coda_statements.write({
-            'state': 'confirm',
-            'date_done': time.strftime("%Y-%m-%d %H:%M:%S")})
         return True
 
     @api.multi
